pyproject.toml
[project]
name = "ai-investor"
version = "0.1.0"
description = "AI Agents for stocks trading for educational porpuses"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiolimiter>=1.2.1",
    "diskcache>=5.6.3",
    "httpx>=0.28.1",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "openai-agents>=0.6.4",
    "pandas>=2.3.3",
    "pre-commit>=4.5.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "types-requests>=2.32.4.20250913",
]

[dependency-groups]
dev = [
    "types-requests>=2.32.4.20250913",
]

[tool.isort]
profile="black"
